backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
faiss-cpu>=1.8.0 
pydantic>=1.10.0
tiktoken>=0.4.0
python-dotenv>=1.0.0
langchain==0.3.24
python-docx>=1.1.1
PyPDF2>=3.0.0
langchain-community==0.3.16
langchain-ollama==0.2.1
PyMuPDF==1.25.5
langgraph==0.3.30